v4

+++ Src/UI_Control/Widget/store.py
@@ -1,10 +1,13 @@
 import sys
 from PyQt5 import QtWidgets
+from PyQt5.QtWidgets import *
+
 from Widget.store_ui import store_ui_widget
 import os
 import cv2
 import shutil
-
+from lib.vis_pose import draw_points_and_skeleton, joints_dict
+import pandas as pd
 
 class Store_Widget(QtWidgets.QWidget):
     def __init__(self, video_name, video_images, person_df):
@@ -19,6 +22,8 @@ class Store_Widget(QtWidgets.QWidget):
         self.video_name = video_name
         self.video_images = video_images
         self.select_id = []
+        
+        self.saving_video = False
         self.add_id_checkbox()
 
     def store(self):
@@ -56,15 +61,19 @@ class Store_Widget(QtWidgets.QWidget):
         return filter_df
     
     def save_datas(self):
+        if self.saving_video:
+            QMessageBox.warning(self, "儲存影片失敗", "請不要多次按下儲存影片按鈕!")
+            return
+        self.saving_video = True
         output_folder = f"../../Db/output/{self.video_name}"
         if os.path.exists(output_folder):
             shutil.rmtree(output_folder)    
         os.makedirs(output_folder)
         
-        # save frame image            
-        for i, frame in enumerate(self.video_images):
-            frame_path = os.path.join(output_folder, f"{i}.jpg")
-            cv2.imwrite(frame_path, frame)
+        # # save frame image            
+        # for i, frame in enumerate(self.video_images):
+        #     frame_path = os.path.join(output_folder, f"{i}.jpg")
+        #     cv2.imwrite(frame_path, frame)
 
        # 将 DataFrame 保存为 JSON 文件
         json_path = os.path.join(output_folder, f"{self.video_name}.json")
@@ -74,7 +83,46 @@ class Store_Widget(QtWidgets.QWidget):
             save_person_df = self.person_df
         else:
             save_person_df = self.filter_person_df()
+
         save_person_df.to_json(json_path, orient='records')
+        video_size = (1280, 720)
+        fps = 30.0
+        save_location = "../../Db/output/" + self.video_name + "/" + self.video_name + ".mp4"
+
+        video_writer = cv2.VideoWriter(save_location, cv2.VideoWriter_fourcc(*'mp4v'), fps, video_size)
+
+        if not video_writer.isOpened():
+            print("error while opening video writer!")
+            return
+        
+        for frame_num, frame in enumerate(self.video_images):
+            curr_person_df = self.obtain_frame_data(frame_num,save_person_df)
+            image = self.draw_image(frame_num,frame, curr_person_df)
+            video_writer.write(image)
+
+        video_writer.release()
+
+        self.saving_video = False
+        QMessageBox.information(self, "儲存影片", "影片儲存完成!")
+            
+    def draw_image(self,frame_num,frame,person_df):    
+        image=frame.copy()
+        if not person_df.empty:
+            image = draw_points_and_skeleton(image, person_df, joints_dict()['coco']['skeleton_links'], 
+                                            points_color_palette='gist_rainbow', skeleton_palette_samples='jet',
+                                            points_palette_samples=10, confidence_threshold=0.3)
+        else:
+            pass
+
+        return image
+    
+    def obtain_frame_data(self,frame_num,person_df):
+        curr_person_df = pd.DataFrame()
+        try :
+            curr_person_df = person_df.loc[(person_df['frame_number'] == frame_num)]
+        except ValueError:
+            print("valueError")
+        return curr_person_df
 
     def cancel(self):
         self.close()
